move jwt to middleware folder and update imports
