+++ middlewares/isAdmin.js
@@ -1,7 +1,7 @@
 const ApiError = require('../helpers/ApiError')
 const { catchAsync } = require('../helpers/catchAsync')
 const httpStatus = require('../helpers/httpStatus')
-const { decodeToken } = require('../services/jwt')
+const { decodeToken } = require('./jwt')
 
 module.exports = {
   isAdmin: catchAsync(async (req, res, next) => {
